refactor: move cookie logic to cookies modules
